improve -Made tests/ directory and move all tests files to this directory.- fix -Fixed import error bug in test files.-
